cinePlayer.cpp progress

+++ include/Camera.h
@@ -189,8 +189,8 @@ struct CamDataInfo {
 	Vector3f _00;   // _00
 	Vector3f _0C;   // _0C
 	Vector3f _18;   // _18
-	u32 _24;        // _24
-	u32 _28;        // _28
+	f32 _24;        // _24
+	f32 _28;        // _28
 	char _2C;       // _2C
 	u32 _30[9];     // _30, sizeof 0x24
 	u32 _54[21];    // _54, sizeof 0x54
